Updated 23/04/2018

+++ Include/ObjNode.h
@@ -17,17 +17,17 @@ public:
 
     void setObjNodeDepth(int objNodeDepth);
 
-    int getObjNodeX();
+    double getObjNodeX();
 
-    void setObjNodeX(int objNodeX);
+    void setObjNodeX(double objNodeX);
 
-    int getObjNodeY();
+    double getObjNodeY();
 
-    void setObjNodeY(int objNodeY);
+    void setObjNodeY(double objNodeY);
 
-    int getObjNodeWid();
+    float getObjNodeWid();
 
-    void setObjNodeWid(int objNodeWid);
+    void setObjNodeWid(float objNodeWid);
 
     int getNumberOfChildren();
 
@@ -40,9 +40,9 @@ public:
 
 private:
     int objNodeDepth;
-    int objNodeX;
-    int objNodeY;
-    int objNodeWid;
+    double objNodeX;
+    double objNodeY;
+    float objNodeWid;
     int numberOfChildren;
     int sonsCount = 0;
     //int ovjNodeHeigth; per ora li facciamo tutti alla stessa altezza
